# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CryptoBoiler API",
    "description": "API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cryptoboiler.duckdns.org:7443",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/telemetry/": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "description": "",
        "parameters": [
          {
            "name": "boiler_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/components/responses/SuccessOK"
          },
          "301": {
            "$ref": "#/components/responses/MovedPermanently",
            "description": "Moved Permanently"
          },
          "302": {
            "$ref": "#/components/responses/Found",
            "description": "Found"
          },
          "304": {
            "$ref": "#/components/responses/NotModified",
            "description": "Not Modified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Bad Request"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "Not Found"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests",
            "description": "Too Many Requests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError",
            "description": "Internal Server Error"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway",
            "description": "Bad Gateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable",
            "description": "Service Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Telemetry"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/components/responses/SuccessOK"
          },
          "301": {
            "$ref": "#/components/responses/MovedPermanently",
            "description": "Moved Permanently"
          },
          "302": {
            "$ref": "#/components/responses/Found",
            "description": "Found"
          },
          "304": {
            "$ref": "#/components/responses/NotModified",
            "description": "Not Modified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Bad Request"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "Not Found"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests",
            "description": "Too Many Requests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError",
            "description": "Internal Server Error"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway",
            "description": "Bad Gateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable",
            "description": "Service Unavailable"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "boiler_id": {
                    "example": "any"
                  },
                  "temperature": {
                    "example": "any"
                  },
                  "pressure": {
                    "example": "any"
                  },
                  "timestamp": {
                    "example": "any"
                  },
                  "hmac_signature": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "responses": {
      "SuccessOK": {
        "description": "Успешное выполнение (Код 200)"
      },
...
      "MovedPermanently": {
        "description": "Ресурс перемещен навсегда (Код 301)"
      },
      "Found": {
        "description": "Ресурс временно перемещен (Код 302)"
      },
      "NotModified": {
        "description": "Ресурс не изменялся с момента последнего запроса (Код 304)"
      },
...
      "BadRequest": {
        "description": "Неверный синтаксис или ошибка валидации (Код 400)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Сервер не понял запрос"
                }
              }
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Превышен лимит запросов (Код 429)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Превышен лимит запросов"
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Сервер временно недоступен (Код 503)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Сервер на обслуживании"
                }
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Запрашиваемый ресурс не найден (Код 404)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Ресурс не найден"
                }
              }
            }
          }
        }
      },
      "BadGateway": {
        "description": "Шлюз получил некорректный ответ от устройства (Код 502)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Котел недоступен"
                }
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Внутренняя ошибка сервера (Код 500)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Внутренняя ошибка сервера"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ]
}